Flytt dupliser knapp til handlinger
